+++ Assets/Prefabs/Character/Player.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4435ca00a4b7ed94ba6a732efa843ae2
+guid: 9f8e21282b3b1884183c7faeaf0b8313
 PrefabImporter:
   externalObjects: {}
   userData:
