+++ Assets/Scripts/Character/ModelController.cs
@@ -18,21 +18,14 @@ public class ModelController : MonoBehaviour
     public GameObject[] m_bodys;
    
     public WeaponController[] m_weapons;
-
     public Dictionary<eWeaponType, WeaponController> dicWeapons;
 
     public Character character;
 
-    public Rig HeadAnim;
-    public Rig HandRighAnim;
-    public Rig HandLeftAnim;
-    public Rig LowerarmLeftAnim;
-    public Rig LowerarmRightAnim;
-
 
     //파츠바디 타입에는 무기를 포함하지않는다.
     public Dictionary<eEuipmentType, List<GameObject>> partsByType = new Dictionary<eEuipmentType, List<GameObject>>();
-    private void InitializeParts()
+    protected virtual void InitializeParts()
     {
         // 배열을 Dictionary에 등록합니다.
         partsByType[eEuipmentType.Head] = new List<GameObject>(m_heads);
@@ -53,34 +46,8 @@ public class ModelController : MonoBehaviour
                 //Debug.LogWarning($"[{partType}] 파츠 목록이 비어있습니다.");
             }
         }
-        dicWeapons = new Dictionary<eWeaponType, WeaponController>();
-        foreach (var weapon in m_weapons)
-        {
-            dicWeapons.Add(weapon.eWeaponType, weapon);
-        }
-
-        ScanForTargets.OnSetLockOnTarget += OnSetLockOnTarget;
     }
 
-    public WeightedTransform posxxx;
-
-    public void OnSetLockOnTarget(Transform Target)
-    {
-
-    }
-
-    private void AimingPosesRigging(Rig weiponAimLayer, Transform Target)
-    {
-
-
-        weiponAimLayer.weight = 1;
-        GameObject child = weiponAimLayer.transform.GetChild(0).gameObject;
-        MultiAimConstraint childCompo =  child.GetComponent<MultiAimConstraint>();
-
-        childCompo.weight = 1;
-        posxxx = childCompo.data.sourceObjects[0];
-        
-    }
 
 
 
@@ -139,24 +106,25 @@ public class ModelController : MonoBehaviour
         }
     }
 
-    public void SetWeaponByIndex(eWeaponType type)
-    {      
-        foreach(var weapon in dicWeapons)
+
+
+    #endregion
+
+
+    public virtual void SetWeaponByIndex(eWeaponType type)
+    {
+        foreach (var weapon in dicWeapons)
         {
             bool isActive = type == weapon.Key;
             dicWeapons[weapon.Key].gameObject.SetActive(isActive);
 
             //활성화된 현제 무기
             if (isActive)
-            {                
+            {
                 character.SetWeaponEffect(dicWeapons[type]);
             }
         }
     }
-
-
-    #endregion
-
 
 
 }
