+++ Assets/Scripts/Character/RigModelController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 90808514f8898504998f87cb37477b25
